[sonar] Explicit cast when loosing precision (C++ examples).

+++ examples/s4u/actor-yield/s4u-actor-yield.cpp
@@ -19,7 +19,7 @@ class yielder {
   long number_of_yields;
 
 public:
-  explicit yielder(std::vector<std::string> args) { number_of_yields = std::stod(args[1]); }
+  explicit yielder(std::vector<std::string> args) { number_of_yields = std::stol(args[1]); }
   void operator()() const
   {
     for (int i = 0; i < number_of_yields; i++)
